Continuing on PPC emulation (Artesyn PM/PPC stuff mostly). Adding an Algor emulation skeleton; implementing some of the basics.
git-svn-id: file:///home/ryo_on/repos/gxemul-repo/gxemul-legacy/trunk@3692 dbbdf1f2-f3a0-11dc-8807-f0e41a653009

+++ src/include/pcireg.h
@@ -1,4 +1,4 @@
-/*  gxemul: $Id: pcireg.h,v 1.5 2005-11-08 11:01:48 debug Exp $  */
+/*  gxemul: $Id: pcireg.h,v 1.6 2005-11-17 13:53:43 debug Exp $  */
 /*	$NetBSD: pcireg.h,v 1.37 2002/03/22 20:03:20 drochner Exp $	*/
 
 #ifndef _DEV_PCI_PCIREG_H_
@@ -71,8 +71,8 @@ typedef u_int16_t pci_product_id_t;
 	    (((id) >> PCI_PRODUCT_SHIFT) & PCI_PRODUCT_MASK)
 
 #define PCI_ID_CODE(vid,pid)					\
-	((((vid) & PCI_VENDOR_MASK) << PCI_VENDOR_SHIFT) |	\
-	 (((pid) & PCI_PRODUCT_MASK) << PCI_PRODUCT_SHIFT))
+    ((((vid) & PCI_VENDOR_MASK) << PCI_VENDOR_SHIFT) |		\
+     (((uint32_t)((pid) & PCI_PRODUCT_MASK)) << PCI_PRODUCT_SHIFT))
 
 /*
  * Command and status register.
